Core v1.3
the core of logic gates has been changes YET again!

this time with some optimization for data usage
now with nodes and each node keeps track of its own value instead of an array of values so now updating a node is much faster and it doesnt sweep over the whole array!